changed folder structure and added some helpers
